increase PointId to uint64_t

+++ include/pdal/pdal_types.hpp
@@ -46,7 +46,7 @@
 namespace pdal
 {
 
-typedef uint32_t PointId;
+typedef uint64_t PointId;
 typedef uint64_t point_count_t;
 typedef std::vector<std::string> StringList;
